[SC-65] refactor : comment, replyComment 패키지 분리 및 코드 정리
